requirements.txt
gradio==4.27.0
google-generativeai>=0.5.2
python-dotenv
